Use real-life, actual, bonafide files instead of trying to mock
I was using mock-fs, which is really nice for tests. However, it doesn't seem to really work for dynamic requires, and is designed to mock calls to the `fs` module. Instead of only running on certain platforms, let's just use real files.
