+++ package.json
@@ -27,7 +27,6 @@
     "@testing-library/react": "^11.1.2",
     "@types/jest": "^26.0.15",
     "@types/lodash": "^4.14.165",
-    "@types/mock-fs": "^4.13.0",
     "@types/react": "^16.9.56",
     "@typescript-eslint/eslint-plugin": "^4.8.1",
     "@typescript-eslint/parser": "^4.8.1",
@@ -39,7 +38,6 @@
     "eslint-plugin-testing-library": "^3.10.0",
     "jest": "^26.6.3",
     "lerna": "^3.22.1",
-    "mock-fs": "^4.13.0",
     "prettier": "^2.1.2",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
