+++ packages/story-utils/src/__tests__/__mock_stories__/bar.stories.tsx
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+
+export default {
+  title: 'Bar',
+  component: () => 'bar',
+};
+
+export const BarStory1 = () => 'bar story 1';
